==> 1_code/02a_Reactome_DB_Queries.ipynb <==
# Extracting Edges from Reactome

Reactome exists in its purest form as a Neo4j data dump.  We will download neo4j, populate it with the reactome dump, then start the server and extract the edges we're interested in.

We will be extracting the following edge types:

- Mappings to external resources
    - Physical Entities and Events to Taxa
    - Physical Entities and Events to GO Biological Process
    - Physical Entities and Events to GO Cell Component
    - Physical Entities and Events to GO Molecular Function
    - Physical Entities and Events to Disease
        - Rxn AssocWith Disease
        - Rxn Disrupted in Disease
        - PW AssocWith Disease
        - PW Disrupted in Disease
    - Taxa to Disease Through intermediary node...
- Mappings between internal Resources
    - Failed Rxn to Rxn
    - Disrupted PW to PW
- Membership Details
    - Preceeding Reaction to Post Reaction
    - Pathway has event Reaction
    - PE input and output of Reactions
    - PE Component of Complex
- Regulation
    - Postive Regulation
    - Negative Regulation
  

import pandas as pd
from pathlib import Path
from data_tools import download

## Download Neo4j Database

dl_loc = Path('../0_data/external/').resolve()
neo4j_url = 'https://neo4j.com/artifact.php?name=neo4j-community-3.5.12-unix.tar.gz'
neo_f_name = neo4j_url.split('=')[-1]

download(neo4j_url, dl_loc.joinpath(neo_f_name))

### Extract neo4j to the parent directory

neo_dir = Path('../').resolve()

import tarfile

with tarfile.open(dl_loc.joinpath(neo_f_name), 'r:gz') as n4j:
    n4j.extractall(neo_dir)
    neo_dir = neo_dir.joinpath(n4j.getnames()[0])

## Download Reactome Neo4j Dump

reactome_db_url = 'https://reactome.org/download/current/reactome.graphdb.tgz'
rct_f_name = reactome_db_url.split('/')[-1]

download(reactome_db_url, dl_loc.joinpath(rct_f_name))

### Extact Reactome to the database directory within neo4j

db_dir = neo_dir.joinpath('data/databases')

with tarfile.open(dl_loc.joinpath(rct_f_name), 'r:gz') as react_db:
    react_db.extractall(db_dir)

## Spin up the neo4j server

import subprocess
import time

command = str(neo_dir.joinpath('bin/neo4j')) + ' start'
subprocess.call(command, shell=True)
# Should take about < 5 seconds to spin up the new neo4j database server
time.sleep(5)

### Authenticate to connect to the server

import neo4j

try:
    # Use Neo4j Default port, user, and password
    driver = neo4j.GraphDatabase.driver('bolt://localhost:7687', auth=("neo4j", "neo4j"))
    session = driver.session()

    # We are required by Neo4j to change the password before any queries can may run
    # This data is neither public nor sensitive, so we will use an insecure password
    session.run("CALL dbms.changePassword('new password')");

except:
    # On subsequent runs, password will be different so this will be sued to authenticate
    # Need to wait a bit before re-auth, otherwise error will be thrown.
    time.sleep(5)
    driver = neo4j.GraphDatabase.driver('bolt://localhost:7687', auth=("neo4j", "new password"))
    session = driver.session();

# Run Database Queries

def query_to_df(query):
    txn = session.run(query)
    return pd.DataFrame(txn.data())

query_results = dict()

## Mappings To External Resources

### Physical Entities and Events to Taxa

q1 = "MATCH (n:Taxon)-[e]-(p:PhysicalEntity) RETURN n.taxId, n.displayName, type(e), p.stId, p.displayName"
pe_tax = query_to_df(q1)
len(pe_tax)

pe_tax.head()

pe_tax.columns = ['tax_id', 'tax_name', 'type', 'pe_id', 'pe_name']
pe_tax['tax_id'] = 'NCBITaxon:' + pe_tax['tax_id']
pe_tax.head(2)

query_results['pe_tax'] = pe_tax

q2 = "MATCH (n:Taxon)-[e]-(p:Event) RETURN n.taxId, n.displayName, type(e), p.stId, p.displayName"
evt_tax = query_to_df(q2)
len(evt_tax)

evt_tax.head()

evt_tax.columns = ['tax_id', 'tax_name', 'type', 'evt_id', 'evt_name']
evt_tax['tax_id'] = 'NCBITaxon:' + evt_tax['tax_id']
evt_tax.head(2)

query_results['evt_tax'] = evt_tax

### Physical Entities and Events to GO Biological Process

q3 = """MATCH (pe:PhysicalEntity)-[e]-(cc:GO_BiologicalProcess) 
        RETURN pe.stId, pe.displayName, type(e), cc.databaseName, cc.accession, cc.displayName"""

pe_bp = query_to_df(q3)
len(pe_bp)

pe_bp.head(2)

No results for Physical Entitiy to Biological Process... 

q4 = """MATCH (pe:Event)-[e]-(bp:GO_BiologicalProcess) 
        RETURN pe.stId, pe.displayName, type(e), bp.databaseName, bp.accession, bp.displayName"""

evt_bp = query_to_df(q4)
len(evt_bp)

evt_bp.head(2)

evt_bp.columns = ['evt_id', 'evt_name', 'type', 'curi', 'go_id', 'go_name']
evt_bp['go_id'] = evt_bp['curi'] + ':' + evt_bp['go_id']
evt_bp.head(2)

query_results['evt_bp'] = evt_bp

### Physical Entities and Events to GO Cell Component

q5 = """MATCH (pe:PhysicalEntity)-[e]-(cc:GO_CellularComponent) 
        RETURN pe.stId, pe.displayName, type(e), cc.databaseName, cc.accession, cc.displayName"""

pe_cc = query_to_df(q5)
len(pe_cc)

pe_cc.head(2)

pe_cc.columns = ['pe_id', 'pe_name', 'type', 'curi', 'go_id', 'go_name']
pe_cc['go_id'] = pe_cc['curi'] + ':' + pe_cc['go_id']
pe_cc.head(2)

query_results['pe_cc'] = pe_cc

q6 = """MATCH (pe:Event)-[e]-(cc:GO_CellularComponent) 
        RETURN pe.stId, pe.displayName, type(e), cc.databaseName, cc.accession, cc.displayName"""

evt_cc = query_to_df(q6)
len(evt_cc)

evt_cc.head(2)

evt_cc.columns = ['evt_id', 'evt_name', 'type', 'curi', 'go_id', 'go_name']
evt_cc['go_id'] = evt_cc['curi'] + ':' + evt_cc['go_id']
evt_cc.head(2)

query_results['evt_cc'] = evt_cc

### Physical Entities and Events to GO Molecular Function

q7 = """MATCH (pe:PhysicalEntity)-[e]-(mf:GO_MolecularFunction) 
        RETURN pe.stId, pe.displayName, type(e), mf.databaseName, mf.accession, mf.displayName"""

pe_mf = query_to_df(q7)
len(pe_mf)

pe_mf.head(2)

q8 = """MATCH (pe:Event)-[e]-(mf:GO_MolecularFunction) 
        RETURN pe.stId, pe.displayName, type(e), mf.databaseName, mf.accession, mf.displayName"""

evt_mf = query_to_df(q8)
len(evt_mf)

evt_mf.head(2)

### Physical Entities and Events to Disease

#### Rxn AssocWith Disease

q9 = """Match (s:Reaction)-[e]-(t:Disease) 
        return s.stId, s.displayName, type(e), t.databaseName, t.identifier, t.displayName """

rxn_aw_dis = query_to_df(q9)
len(rxn_aw_dis)

rxn_aw_dis.head(2)

rxn_aw_dis.columns = ['rxn_id', 'rxn_name', 'type', 'curi', 'do_id', 'do_name']
rxn_aw_dis['do_id'] = rxn_aw_dis['curi'] + ':' + rxn_aw_dis['do_id']
rxn_aw_dis['type'] = 'associated_with'
rxn_aw_dis.head(2)

query_results['rxn_aw_dis'] = rxn_aw_dis

####    Rxn Disrupted in Disease

q10 = """Match (s:Reaction)-[e]-(t:FailedReaction)-[e1]-(t1:Disease) 
       return s.stId, s.displayName, type(e), t.displayName, t.stId, type(e1), 
       t1.databaseName, t1.identifier, t1.displayName"""

rxn_di_dis = query_to_df(q10)
len(rxn_di_dis)

rxn_di_dis.head(2)

rxn_di_dis.columns = ['rxn_id', 'rxn_name', 'type', 'drxn_id', 'drxn_name', 'type2', 'curi', 'do_id', 'do_name']
rxn_di_dis['do_id'] = rxn_di_dis['curi'] + ':' + rxn_di_dis['do_id']
rxn_di_dis['type'] = 'disrupted_in'
rxn_di_dis.head(2)

query_results['rxn_di_dis'] = rxn_di_dis

#### PW AssocWith Disease

q11 = """Match (s:Pathway)-[e]-(t:Disease) 
        return s.stId, s.displayName, type(e), t.databaseName, t.identifier, t.displayName"""

pw_aw_dis = query_to_df(q11)
len(pw_aw_dis)

pw_aw_dis.head(2)

pw_aw_dis.columns = ['pw_id', 'pw_name', 'type', 'curi', 'do_id', 'do_name']
pw_aw_dis['do_id'] = pw_aw_dis['curi'] + ':' + pw_aw_dis['do_id']
pw_aw_dis['type'] = 'associated_with'
pw_aw_dis.head(2)

query_results['pw_aw_dis'] = pw_aw_dis

#### PW Disrupted in Disease

q12 = """MATCH (pw:Pathway)<-[e:normalPathway]-(dpw:Pathway)-[e1]-(d:Disease) 
         RETURN pw.stId, pw.displayName, type(e), dpw.stId, dpw.displayName, type(e1), 
         d.databaseName, d.identifier, d.displayName"""

pw_di_dis = query_to_df(q12)
len(pw_di_dis)

pw_di_dis.head(2)

pw_di_dis.columns = ['pw_id', 'pw_name', 'type', 'dpw_id', 'dpw_name', 'type2', 'curi', 'do_id', 'do_name']
pw_di_dis['do_id'] = pw_di_dis['curi'] + ':' + pw_di_dis['do_id']
pw_di_dis['type'] = 'disrupted_in'
pw_di_dis.head(2)

query_results['pw_di_dis'] = pw_di_dis

### Taxa to Disease Trough intermediary node

q13 = """MATCH (n:Taxon)-[e1]-(p)-[e2]-(d:Disease) WHERE n.taxId <> '9606' 
         RETURN DISTINCT n.taxId, n.displayName, d.databaseName, d.identifier, d.displayName"""

tax_dis = query_to_df(q13)
len(tax_dis)

tax_dis.head(2)

tax_dis.columns = ['tax_id', 'tax_name', 'curi', 'do_id', 'do_name']
tax_dis['tax_id'] = 'NCBITaxon:' + tax_dis['tax_id']
tax_dis['do_id'] = tax_dis['curi'] + ':' + tax_dis['do_id']
tax_dis.head(5)

query_results['tax_dis'] = tax_dis

## Mappings between internal Resources

### Failed Rxn to Rxn

q14 = """MATCH (fr:FailedReaction)-[e]-(r:Reaction) 
         RETURN fr.stId, fr.displayName, type(e), r.stId, r.displayName"""
rxn_frxn = query_to_df(q14)
len(rxn_frxn)

rxn_frxn.head(2)

rxn_frxn.columns = ['frxn_id', 'frxn_name', 'type', 'rxn_id', 'rxn_name']
rxn_frxn.head()

query_results['rxn_frxn'] = rxn_frxn

### Disrupted PW to PW

q15 = """MATCH (dpw:Pathway)-[e:normalPathway]->(pw:Pathway) 
         RETURN dpw.stId, dpw.displayName, type(e), pw.stId, pw.displayName"""

pw_dpw = query_to_df(q15)
len(pw_dpw)

pw_dpw.head(2)

pw_dpw.columns = ['dpw_id', 'dpw_name', 'type', 'pw_id', 'pw_name']
pw_dpw.tail()

query_results['pw_dpw'] = pw_dpw

## Membership Details

### Preceeding Reaction to Post Reaction

q16 = """MATCH (rx:Reaction)-[e:precedingEvent]->(rx2:Reaction) 
         RETURN rx.stId, rx.displayName, type(e), rx2.stId, rx2.displayName"""

rxn_p_rxn = query_to_df(q16)
len(rxn_p_rxn)

rxn_p_rxn.head(2)

rxn_p_rxn.columns = ['af_rxn_id', 'af_rxn_name', 'type', 'b4_rxn_id', 'b4_rxn_name']
rxn_p_rxn.drop_duplicates(subset=['af_rxn_name', 'b4_rxn_name']).head()

query_results['rxn_p_rxn'] = rxn_p_rxn

### Pathway has event Reaction

q17 = """MATCH (pw:Pathway)-[e]-(r:Reaction) 
         RETURN pw.stId, pw.displayName, type(e), r.stId, r.displayName"""

pw_rxn = query_to_df(q17)
len(pw_rxn)

pw_rxn.head(2)

pw_rxn.columns = ['pw_id', 'pw_name', 'type', 'rxn_id', 'rxn_name']
pw_rxn.head()

query_results['pw_rxn'] = pw_rxn

### Physical Entity input and output of Reactions

These are covered by the flat file download, however those are missing the input and output data for the physical entities, so they're all `PhysicalEntity part_of Reaction` edges.

q18 = """Match (s:Reaction)-[e]-(t:PhysicalEntity) 
         return s.stId, s.displayName, type(e), t.stId, t.displayName"""

rxn_pe = query_to_df(q18)
len(rxn_pe)

rxn_pe.head(2)

rxn_pe.columns = ['rxn_id', 'rxn_name', 'type', 'pe_id', 'pe_name']
rxn_pe['type'].value_counts()

Mostly interested in `input` and `output` and may drop the other types later...

query_results['rxn_pe'] = rxn_pe

### Physical Entity component of Complexes

Complexes are an internal neo4j type. However many are made up of items that have external identifiers, like Drugs, and Proteins. We would like to extract these relationships for future use. 

q19 = """MATCH (n1:Complex)-[e:hasComponent]-(n2:PhysicalEntity) 
        WHERE n2:Drug OR n2:GenomeEncodedEntity OR n2:SimpleEntity 
        RETURN DISTINCT n1.stId, n1.displayName, n2.stId, n2.displayName, labels(n2)"""

cplx_pe = query_to_df(q19)
len(cplx_pe)

cplx_pe.head(2)

cplx_pe.columns = ['cplx_id', 'cplx_name', 'pe_id', 'pe_name', 'pe_label']
cplx_pe['pe_label'].astype(str).value_counts()

query_results['cplx_pe'] = cplx_pe

## Regulation

q20 = """MATCH (n1)<-[e:regulator]-(r)<-[e2:regulatedBy]-(n2) 
         RETURN DISTINCT n1.stId, n1.displayName, labels(n1), 
         labels(r), n2.stId, n2.displayName, labels(n2)"""

reg_edge = query_to_df(q20)
len(reg_edge)

reg_edge.head(2)

reg_edge.columns = ['n1_id', 'n1_name', 'n1_label', 'type', 'n2_id', 'n2_name', 'n2_label']
reg_edge.head()

pos_idx = reg_edge['type'].apply(lambda x: 'PositiveRegulation' in x)
neg_idx = reg_edge['type'].apply(lambda x: 'NegativeRegulation' in x)

reg_edge.loc[pos_idx[pos_idx].index, 'type'] = 'PositiveRegulation'
reg_edge.loc[neg_idx[neg_idx].index, 'type'] = 'NegativeRegulation'
reg_edge['type'].value_counts()

Guessing that is almost always a PhysicalEntity that regulates an Event... lets see if thats true

reg_edge['n1_label'].apply(lambda x: 'PhysicalEntity' in x).sum() == len(reg_edge)

reg_edge['n2_label'].apply(lambda x: 'Event' in x).sum() == len(reg_edge)

reg_edge['n1_label'] = 'PhysicalEntity'
reg_edge['n2_label'] = 'Event'

reg_edge.columns = ['pe_id', 'pe_name', 'pe_label', 'type', 'evt_id', 'evt_name', 'evt_label']

reg_edge.head(5)

query_results['reg_edge'] = reg_edge

# Save and exit

nb_name = '02a_Reactome_DB_Queries'
out_dir = Path('../2_pipeline/').joinpath(nb_name, 'out').resolve()

# Make the output directory if doesn't already exist
out_dir.mkdir(parents=True, exist_ok=True)

for name, df in query_results.items():
    df.to_csv(out_dir.joinpath(name+'.csv'), index=False)

#### Shutdown the Neo4j server

subprocess.call(command.replace('start', 'stop'), shell=True)
