==> 1_code/05c_Computing_GO_Regulation_Edges.ipynb <==
# GO Nodes into Edges

GO Biological Process nodes sometimes have to do with regulation...

E.G. GO:0045861	negative regulation of proteolysis	

Any Term that links to this go term essentially has an edge:
    negatively_regulates -> GO:0006508	proteolysis

Therefore we will determine these regulation nodes and convert them to edges.

import pandas as pd
from pathlib import Path
from data_tools.df_processing import regularize_colnames, expand_col_on_char
from data_tools import graphs as gt
from data_tools import obo_processing as ot

%matplotlib inline
import seaborn as sns
import matplotlib.pyplot as plt

load_dir = Path('../2_pipeline/00_download_data/out/').resolve()
network_dir = load_dir.parent.parent.joinpath('05b_Other_Ontologies/out')

this_name = '05c_Computing_GO_Regulation_Edges'
out_dir = Path('../2_pipeline/').joinpath(this_name).joinpath('out').resolve()
out_dir.mkdir(parents=True, exist_ok=True)

all_nodes = pd.read_csv(network_dir.joinpath('nodes.csv'), dtype=str)
edges = pd.read_csv(network_dir.joinpath('edges.csv'), dtype=str)

all_node_ids = all_nodes['id'].values
new_nodes = []
new_edges = []

all_nodes.head(2)

# GO regulation...

go_terms = all_nodes.query('label in {}'.format(['Molecular Function', 'Cellular Component', 'Biological Process']))
len(go_terms)

reg_terms = go_terms['name'].str.contains('regulation of')
reg_terms.sum()

reg_ids = go_terms[reg_terms]['id'].values

## We'll get the regulation Edges from go.obo

Need some functions to parse the obo file...

go_rels = ot.get_ontology_edges(load_dir.joinpath('go.obo'))

go_rels.head(10)

go_rels['rel_type'].value_counts()

go_rels.query('rel_type == "part_of"').sample(5)

reg_rels = ['regulates', 'negatively_regulates', 'positively_regulates']
reg_res = go_rels.query('rel_type in @reg_rels')

reg_ids = reg_res['src_id'].values

combo = gt.combine_nodes_and_edges(all_nodes, edges)
combo_cols = ['start_id', 'start_name', 'start_label', 'type', 'end_id', 'end_name', 'end_label']

no_map_terms = go_terms[reg_terms].query('id not in @reg_ids')['id'].values
print('Number of edges containing Regulation terms that cannot be converted to an edge: {:,}'.format(
            len(combo.query('start_id in @no_map_terms and end_label == "Disease"'))))
combo.query('start_id in @no_map_terms and end_label == "Disease"')[combo_cols].sample(5)

map_terms = go_terms.query('id in @reg_ids')['id'].values
print('Number of edges containing Regulation terms that can become an edge: {:,}'.format(
    len(combo.query('start_id in @map_terms and end_label == "Disease"'))))
combo.query('start_id in @map_terms and end_label == "Disease"')[combo_cols].sample(5)

tgt_ids = reg_res['tgt_id'].values
map_terms1 = go_terms.query('id in @tgt_ids')['id'].values
print(len(combo.query('start_id in @map_terms1 and end_label == "Disease"')))
combo.query('start_id in @map_terms1 and end_label == "Disease"')[combo_cols].sample(5)

# Regulations of regulation.... not sure what to do here but just drop
print(len(reg_res[reg_res['tgt_name'].str.contains('regulation of')]))
reg_res[reg_res['tgt_name'].str.contains('regulation of')].sample(5)

# Things without 'regulation of' that appearantly regulate...
print(len(reg_res[~reg_res['src_name'].str.contains('regulation of')]))
reg_res[~reg_res['src_name'].str.contains('regulation of')].sample(5)

edges.columns

combo_cols += ['source', 'dsrc_type', 'comp_type', 'p_val', 'adj_p', 'license']
mrg = pd.merge(combo[combo_cols], reg_res, how='left', left_on='end_id', right_on='src_id')

'{:,}'.format(len(mrg))

reg_edges = mrg.dropna(subset=['tgt_id'])

reg_edges.head(2)

id_to_label = all_nodes.set_index('id')['label'].to_dict()

reg_edges = reg_edges.drop(['type', 'end_id', 'end_label', 'end_name'], axis=1)
reg_edges['end_label'] = reg_edges['tgt_id'].map(id_to_label)
reg_edges = reg_edges.rename(columns={'tgt_id':'end_id', 'tgt_name': 'end_name', 'rel_type': 'type'})
reg_edges.head(2)

reg_edges = reg_edges[combo_cols].drop(['start_name', 'start_label', 'end_name', 'end_label'], axis=1).reset_index(drop=True)
reg_edges['dsrc_type'] = 'computed'
reg_edges['comp_type'] = 'punning'
reg_edges.head(2)

reg_edges['source'].value_counts()

# Putting it all together

reg_edges.head(2)

reg_edges['type'].value_counts()

print('Total number of new edges: {:,}'.format(len(reg_edges)))
print('Number of unique new edges: {:,}'.format(len(reg_edges.drop_duplicates(subset=['start_id', 'end_id', 'type']))))

Why are so many edges duplicated? what kind are they?

ix = reg_edges.duplicated(keep=False)
reg_edges[ix].sort_values(['start_id', 'end_id', 'type']).head(10)

reg_edges[ix]['type'].value_counts()

dup_types = reg_edges[ix]['type'].unique()

dup_dfs = []

for dt in dup_types:
    dup_dfs.append(reg_edges[ix].query('type == @dt').sort_values(['start_id', 'end_id', 'type']).head(6))
    
pd.concat(dup_dfs)

I see no reason for any of these duplications.... We'll just do a simple 'drop_duplicates' rather than merging any columns

reg_edges = reg_edges.drop_duplicates(subset=['start_id', 'end_id', 'type']).copy()

len(reg_edges)

all_node_ids = set(all_nodes['id']) 

print(len(reg_edges))
reg_edges_filt = reg_edges.query('start_id in @all_node_ids and end_id in @all_node_ids')
print(len(reg_edges_filt))

all_edges = pd.concat([edges, reg_edges_filt], sort=False, ignore_index=True)
all_edges.head(2)

all_edge_ids = all_edges[['start_id', 'end_id']].stack()
filt_nodes = all_nodes.query('id in @all_edge_ids')
len(filt_nodes)

# Save to Disk

all_nodes.to_csv(out_dir.joinpath('nodes.csv'), index=False)
filt_nodes.to_csv(out_dir.joinpath('nodes_filt.csv'), index=False)

all_edges.to_csv(out_dir.joinpath('edges.csv'), index=False)
