==> 1_code/06_HPO_Annotations.ipynb <==
%matplotlib inline

import pandas as pd
from pathlib import Path
from biothings_client import get_client

from data_tools import graphs as gt
from data_tools.wiki import get_curi_xrefs
from data_tools import obo_processing as ot
from data_tools.df_processing import expand_col_on_char, expand_split_col, combine_group_cols_on_char

data_dir = Path('../2_pipeline/00_download_data/out/').resolve()
this_name = '06_HPO_Annotations'
out_dir = Path('../2_pipeline/').joinpath(this_name).joinpath('out').resolve()

if not out_dir.exists():
    out_dir.mkdir(parents=True)

prev_nodes = pd.read_csv('../2_pipeline/05c_Computing_GO_Regulation_Edges/out/nodes.csv', dtype=str)
prev_edges = pd.read_csv('../2_pipeline/05c_Computing_GO_Regulation_Edges/out/edges.csv', dtype=str)

## Get previous phenotypes and xrefs and see what we can match to HPO

pheno_nodes = prev_nodes.query('label == "Phenotype"')

nw_to_xref = expand_col_on_char(pheno_nodes, 'xrefs', '|', False)[['id', 'xrefs']]

nw_to_xref['nw_curi'] = nw_to_xref['id'].apply(lambda s: s.split(':')[0])
nw_to_xref['xref_curi'] = nw_to_xref['xrefs'].apply(lambda s: s.split(':')[0] if type(s) == str else s)

nw_to_xref[['nw_curi', 'xref_curi']].stack().unique()

hp_nodes = ot.get_ontology_nodes(data_dir.joinpath('hpo.obo'))
hp_struct = ot.get_ontology_edges(data_dir.joinpath('hpo.obo'))

hpo_to_xref = expand_col_on_char(hp_nodes[['id', 'xref']], 'xref', '|', False)
hpo_to_xref['xref_curi'] = hpo_to_xref['xref'].apply(lambda s: s.split(':')[0] if type(s) == str else s)

hpo_to_xref['xref_curi'].unique()

# Fix the curis so they're common between our network and HPO
hpo_to_xref['xref'] = (hpo_to_xref['xref'].str.replace('MSH:', 'MESH:')
                                          .str.replace('SNOMEDCT_US:', 'SNOMED:')
                                          .str.replace('MeSH:', 'MESH:'))
hpo_to_xref['xref_curi'] = hpo_to_xref['xref'].apply(lambda s: s.split(':')[0] if type(s) == str else s)


common_curi = set(nw_to_xref[['nw_curi', 'xref_curi']].stack().unique()) & set(hpo_to_xref['xref_curi'].unique())
common_curi

hpo_in_nw = nw_to_xref.query('nw_curi == "HP"')['id'].unique()

new_hpo = hpo_to_xref.query('id not in @hpo_in_nw and xref_curi in @common_curi')
len(new_hpo), new_hpo['id'].nunique()

Look at what terms we can re-align with HPO IDs thanks to these xrefs

match_ids = nw_to_xref.query('nw_curi != "HP" and nw_curi in @common_curi or xref_curi in @common_curi')[['id', 'xrefs']].stack().unique()

found_pheno = new_hpo.query('xref in @match_ids').rename(columns={'id': 'hp_id'})
found_pheno['xref_curi'].value_counts()

About 10% were found to already be in the network, mostly via MESH cross-references.

pheno_reid_map = found_pheno.merge(nw_to_xref.query('nw_curi != "HP"').rename(columns={'id': 'nw_id'}),
                                   left_on='xref', right_on='xrefs')[['hp_id', 'nw_id']]

len(prev_nodes)

changed_ids = pheno_reid_map['hp_id'].unique()

prev_edges = gt.re_id_edges(prev_edges, pheno_reid_map, 'nw_id', 'hp_id')
prev_nodes = gt.re_id_nodes(prev_nodes, pheno_reid_map, 'nw_id', 'hp_id', False)

prev_nodes.query('id in @changed_ids')

# get_official hpo names into the network
hpid_to_name = hp_nodes.set_index('id')['name'].to_dict()

prev_nodes['name'] = prev_nodes['id'].map(hpid_to_name).fillna(prev_nodes['name'])

prev_nodes.query('id in @changed_ids')

## Bring in the Annotations

tab_cols = ['db', 'db_object_id', 'db_name', 'qualifier', 'hp_id', 'db_reference', 'evidence_code', 'onset_modifier', 
            'frequency', 'sex', 'modifier', 'aspect', 'date_created', 'assigned_by']

hp_anno = pd.read_csv(data_dir.joinpath('phenotype_annotation.tab'), header=None, 
                       sep='\t', dtype=str, names=tab_cols)
hp_gene = pd.read_csv(data_dir.joinpath('ALL_SOURCES_ALL_FREQUENCIES_genes_to_phenotype.txt'), 
                       sep='\t', header=None, comment='#', dtype=str,
                       names=['gene_id', 'gene_symbol', 'hp_name', 'hp_id'])

hp_name_map = hp_nodes.set_index('id')['name'].to_dict()
hp_anno['hp_name'] = hp_anno['hp_id'].map(hp_name_map)

len(hp_anno), len(hp_gene)

hp_anno.head(2)

hp_gene.head(2)

## Examine Genes

Genes are using entrez gene_id, so looking for overlap to the current network should be stratightforward

hp_gene['gene_id'] = 'NCBIGene:' + hp_gene['gene_id']
hp_genes = hp_gene['gene_id'].unique()

prev_nodes.query('id in @hp_genes')

hp_gene['gene_id'].nunique()

node_ids = set(prev_nodes['id'])
set(hp_gene['gene_id']) - node_ids

Only a very small number are missing... Don't think we'll add them, but let's at least look at what they

mg = get_client('gene')
mg.getgenes([g.split(':')[-1] for g in list(set(hp_gene['gene_id']) - node_ids)], fields=['name', 'symbol', 'uniprot'])

hp_to_gene = hp_gene.query('gene_id in @node_ids')
len(hp_to_gene)

## Now HP to Disease Links

HP has OMIM ids for diseases... We also have some OMIM Xrefs from WikiData that we can use to merge these concepts.

hp_anno.query('db == "OMIM"')['db_object_id'].nunique()

hp_anno['disease_id'] = hp_anno['db'] + ':' + hp_anno['db_object_id']

prev_nodes.query('label == "Disease"')['id'].nunique()

dis_to_omim = get_curi_xrefs(prev_nodes.query('label == "Disease"'), 'OMIM')
print("Number of OMIM not-mappable to our NW: {:,}".format(len(
    set(hp_anno.query('db == "OMIM"')['disease_id']) - set(dis_to_omim['xrefs']))))

found_diseases = set(hp_anno.query('db == "OMIM"')['disease_id']) & set(dis_to_omim['xrefs'])
hp_to_dis = hp_anno.query('disease_id in @found_diseases').copy()

print("Number of OMIM mapped to our NW: {:,}".format(hp_to_dis['disease_id'].nunique()))
print("Number of annotations: {:,}".format(len(hp_to_dis)))

hp_to_dis['evidence_code'].value_counts()

hp_to_dis = hp_to_dis.sort_values(['hp_id', 'disease_id', 'evidence_code']).drop_duplicates(subset=['disease_id', 'hp_id'])
hp_to_dis.head(4)

hp_to_dis = hp_to_dis.merge(dis_to_omim, how='left', left_on='disease_id', right_on='xrefs').rename(columns={'id': 'nw_id'})

hp_to_dis.head(2)[['nw_id', 'hp_id', 'evidence_code']]

## Figure out which new phenotypes need to be added

keep_hp_ids = set(prev_nodes.query('label == "Phenotype"')['id']) | (set(hp_to_dis['hp_id']) & set(hp_to_gene['hp_id']))

old_hp_ids = prev_nodes.query('label == "Phenotype"')['id']

keep_hp_nodes = hp_nodes.query('id in @keep_hp_ids and id not in @old_hp_ids')[['id', 'name', 'xref']]

# Standardize column name to what we have in the network
keep_hp_nodes = keep_hp_nodes.rename(columns={'xref': 'xrefs'})

# Standarize the curis to what we have in the network
keep_hp_nodes['xrefs'] = (keep_hp_nodes['xrefs'].str.replace('MSH:', 'MESH:')
                                                .str.replace('SNOMEDCT_US:', 'SNOMED:')
                                                .str.replace('MeSH:', 'MESH:'))
# Get rid of xrefs to DBs we don't already look out for
keep_hp_nodes['xrefs'] = keep_hp_nodes['xrefs'].apply(lambda s: '|'.join(sorted([x for x in str(s).split('|') 
                                                                          if x.split(':')[0] in common_curi])))
# Logic above destroys NaNs, so fix that.
keep_hp_nodes['xrefs'] = keep_hp_nodes['xrefs'].replace('', float('nan'))
keep_hp_nodes['label'] = 'Phenotype'
keep_hp_nodes.head()

## Aggregate the new edges and save to disk

hp_to_dis.head(2)

hp_to_dis_edges = hp_to_dis.query('hp_id in @keep_hp_ids').rename(columns={'nw_id': 'start_id', 'hp_id': 'end_id'})

hp_to_dis_edges['dsrc_type'] = hp_to_dis_edges['evidence_code'].apply(lambda e: 'computed' if e == 'IEA' else 'curated')

comp_idx = hp_to_dis_edges[hp_to_dis_edges['dsrc_type'] == 'computed'].index
hp_to_dis_edges.loc[comp_idx, 'comp_type'] = hp_to_dis_edges.loc[comp_idx, 'evidence_code']

hp_to_dis_edges['type'] = 'presents'

keep_edge_cols = ['start_id', 'end_id', 'type', 'dsrc_type', 'comp_type']

hp_to_dis_edges[keep_edge_cols].head(2)

hp_to_gene_edges = hp_to_gene.query('hp_id in @keep_hp_ids').rename(columns={'hp_id': 'start_id', 'gene_id': 'end_id'})
hp_to_gene_edges['type'] = 'associated_with'
hp_to_gene_edges['dsrc_type'] = 'computed'

# from https://hpo.jax.org/app/download/annotation
# ""All phenotype terms associated with any disease that is associated with 
# variants in a gene are assigned to that gene in this file.""
# This sounds like punning to me: (A->B->C becomes A->C)
hp_to_gene_edges['comp_type'] = 'punning'

hp_to_gene_edges[keep_edge_cols].head(2)

all_hp_edges = pd.concat([hp_to_dis_edges[keep_edge_cols], 
                       hp_to_gene_edges[keep_edge_cols]], ignore_index=True)
all_hp_edges.head(2)

all_hp_edges['source'] = 'Human Phenotype Ontology'
all_hp_edges['license'] = 'custom cite no-derivatives'

edges_out = pd.concat([prev_edges, all_hp_edges], sort=False, ignore_index=True)

len(edges_out)

edges_out = combine_group_cols_on_char(edges_out, ['start_id', 'end_id', 'type'], sort=True, prog=False)
len(edges_out)

nodes_out = pd.concat([prev_nodes, keep_hp_nodes], sort=False, ignore_index=True)

all_edge_ids = edges_out[['start_id', 'end_id']].stack().unique()

nodes_filt = nodes_out.query('id in @all_edge_ids').copy()

edges_out.duplicated(subset=['start_id', 'end_id', 'type']).sum()

edges_out.to_csv(out_dir.joinpath('edges.csv'), index=False)
nodes_out.to_csv(out_dir.joinpath('nodes.csv'), index=False)
nodes_filt.to_csv(out_dir.joinpath('nodes_filt.csv'), index=False)

# Get a handle on these metapath semmantics

id_to_label = nodes_filt.set_index('id')['label'].to_dict()
#id_to_label = {**{k: 'Compound' for k in id_to_network['new_id']}, **id_to_label}


edges_out['src_label'] = edges_out['start_id'].map(id_to_label)
edges_out['tgt_label'] = edges_out['end_id'].map(id_to_label)

meta_edges = []
for row in edges_out.itertuples():
    meta_edges.append((row.src_label, row.type, row.tgt_label))
meta_edges = pd.Series(meta_edges)
meta_edges.value_counts()

meta_edges = meta_edges.value_counts().to_frame().reset_index()
meta_edges['start_type'] = meta_edges['index'].apply(lambda e: e[0])
meta_edges['rel'] = meta_edges['index'].apply(lambda e: e[1])
meta_edges['end_type'] = meta_edges['index'].apply(lambda e: e[2])
meta_edges = meta_edges.drop(['index'], axis=1).rename(columns={0: 'counts'})
meta_edges = meta_edges.sort_values(['start_type', 'end_type', 'counts'], ascending=[True, True, False])
meta_edges[['start_type', 'rel', 'end_type', 'counts']].to_csv(out_dir.joinpath('meta_edges.csv'), index=False)
