==> 1_code/10a_Disease_Pheno_Cleanup_parse_owls.ipynb <==
# Cleaning up Disease and Phenotype

Diseases and phenotypes have come from different data souces and have different data types.

import pandas as pd
from pathlib import Path
from data_tools import df_processing as dfp

import rdflib
import ontospy

this_name = '10a_Disease_Pheno_Cleanup_parse_owls'
out_dir = Path('../2_pipeline').resolve().joinpath(this_name, 'out')

out_dir.mkdir(parents=True, exist_ok=True)

def uri_to_id(uri):
    if '#' in uri:
        return uri.split('#')[-1]
    return uri.split('/')[-1].replace('_', ':')

def get_class_id(entity):
    return uri_to_id(entity.uri.toPython())

def get_class_name(entity):
    name = entity.bestLabel()
    if type(name) == str:
        return name
    return name.toPython()

def get_property_objects(entity, property_uri):
    out = []
    for t in entity.triples:
        if t[1].toPython() == property_uri:
            obj = t[2]
            if type(obj) == rdflib.term.BNode:
                obj = t[0]
                if type(obj) == rdflib.term.BNode:
                    continue
            
            obj = obj.toPython()
            
            if obj.startswith('http'):
                obj = uri_to_id(obj)
            
            out.append(obj)
    return out

def get_concept_info(entity):
    out = {}
    
    out['id'] = get_class_id(entity)
    out['name'] = get_class_name(entity)
    
    out['synonyms'] = get_property_objects(entity, 'http://www.geneontology.org/formats/oboInOwl#hasExactSynonym')
    out['xrefs'] = get_property_objects(entity, 'http://www.geneontology.org/formats/oboInOwl#hasDbXref')
    out['alt_ids'] = get_property_objects(entity, 'http://www.geneontology.org/formats/oboInOwl#hasAlternativeId')
    out['subsets'] = get_property_objects(entity, 'http://www.geneontology.org/formats/oboInOwl#inSubset')
    
    out = {k: v if v != [] else float('nan') for k, v in out.items()}
    
    
    return out
    

def get_relationship_info(entity):
    out = {}
    
    subjects = get_class_id(entity)
    objects = get_property_objects(entity, 'http://www.w3.org/2002/07/owl#someValuesFrom')
    
    out['sbjs'] = subjects
    out['objs'] = objects
    
    out = {k: v if v != [] else float('nan') for k, v in out.items()}

    
    return out

## Parse Diseases

%%time
do = ontospy.Ontospy('http://purl.obolibrary.org/obo/doid.owl')

dis_info_json = list()
dis_rel_json = list()

for e in do.all_classes:
    dis_info_json.append(get_concept_info(e))
    dis_rel_json.append(get_relationship_info(e))

dis_info_df = pd.DataFrame(dis_info_json)

dfp.expand_df_on_col(dis_info_df, 'xrefs').dropna(subset=['xrefs'])['xrefs'].apply(lambda s: s.split(':')[0]).value_counts()

for col in ['synonyms', 'xrefs', 'alt_ids', 'subsets']:
    dis_info_df[col] = dis_info_df[col].apply(lambda s: dfp.char_combine_iter(s) if type(s) != float else s)

dis_info_df.to_csv(out_dir.joinpath('DO_node_info.csv'), index=False)

dis_rel_df = pd.DataFrame(dis_rel_json)

dis_rel_df = dis_rel_df.dropna(subset=['objs'])
dis_rel_df = dfp.expand_df_on_col(dis_rel_df, 'objs')
dis_rel_df

dis_rel_df['objs'].apply(lambda s: s.split(':')[0]).value_counts().head(50)

dis_rel_df.to_csv(out_dir.joinpath('DO_edge_info.csv'), index=False)

## Parse Phenotypes

%%time
hp = ontospy.Ontospy('http://purl.obolibrary.org/obo/hp.owl')

pheno_info_json = list()
pheno_rel_json = list()

for e in hp.all_classes:
    pheno_info_json.append(get_concept_info(e))
    pheno_rel_json.append(get_relationship_info(e))

pheno_info_df = pd.DataFrame(pheno_info_json)

pheno_info_df.sample(10)

uri_col = 'id'
dfp.expand_df_on_col(pheno_info_df, uri_col).dropna(subset=[uri_col])[uri_col].apply(lambda s: s.split(':')[0]).value_counts().head(15)

uri_col = 'xrefs'
dfp.expand_df_on_col(pheno_info_df, uri_col).dropna(subset=[uri_col])[uri_col].apply(lambda s: s.split(':')[0]).value_counts().head(15)

hpo_idx = pheno_info_df['id'].str.startswith('HP:')
pheno_info_df[hpo_idx].sample(10)

uri_col = 'xrefs'
dfp.expand_df_on_col(pheno_info_df[hpo_idx], uri_col).dropna(subset=[uri_col])[uri_col].apply(lambda s: s.split(':')[0]).value_counts()

for col in ['synonyms', 'xrefs', 'alt_ids', 'subsets']:
    pheno_info_df[col] = pheno_info_df[col].apply(lambda s: dfp.char_combine_iter(s) if type(s) != float else s)

pheno_info_df.to_csv(out_dir.joinpath('HP_node_info.csv'), index=False)

pheno_rel_df = pd.DataFrame(pheno_rel_json)

pheno_rel_df = pheno_rel_df.dropna(subset=['objs'])

pheno_rel_df = dfp.expand_df_on_col(pheno_rel_df, 'objs')
pheno_rel_df

pheno_rel_df['objs'].apply(lambda s: s.split(':')[0]).value_counts().head(50)

hpo_idx = pheno_rel_df['sbjs'].str.startswith('HP:') |  pheno_rel_df['objs'].str.startswith('HP:')

pheno_rel_df[hpo_idx]['objs'].apply(lambda s: s.split(':')[0]).value_counts().head(50)

pheno_rel_df[hpo_idx]['sbjs'].apply(lambda s: s.split(':')[0]).value_counts().head(50)

def get_uri_edges(curi, edges):
    idx = edges['sbjs'].str.startswith(curi+':') | edges['objs'].str.startswith(curi+':')
    return edges[idx]

get_uri_edges('NCBITaxon', pheno_rel_df)['sbjs'].apply(lambda s: s.split(':')[0]).value_counts().head(50)

get_uri_edges('CL', get_uri_edges('NCBITaxon', pheno_rel_df)).head(50)

len(pheno_rel_df[hpo_idx])

pheno_rel_df.to_csv(out_dir.joinpath('HP_edge_info.csv'), index=False)

## Parse Symptoms

symp = ontospy.Ontospy('http://purl.obolibrary.org/obo/symp.owl')

symp_info_json = list()
symp_rel_json = list()

for e in symp.all_classes:
    symp_info_json.append(get_concept_info(e))
    symp_rel_json.append(get_relationship_info(e))

pd.DataFrame(symp_info_json).dropna(subset=['xrefs'])

symp_df = pd.DataFrame(symp_info_json)

len(symp_df)

for col in ['synonyms', 'xrefs', 'alt_ids', 'subsets']:
    symp_df[col] = symp_df[col].apply(lambda s: dfp.char_combine_iter(s) if type(s) != float else s)

symp_df = symp_df.drop('subsets', axis=1)
symp_df.to_csv(out_dir.joinpath('SYMP_node_info.csv'), index=False)

pd.DataFrame(symp_rel_json).dropna(subset=['objs'])

No relationships to save...
