==> 1_code/10b_Diseae_Phenotype_Cleanup_Preliminary_DOID_cleaning.ipynb <==
# Cleaning up some DOID info

It was noticed that some disease ontology IDs have a cross-reference to another disease ontology ID... they should all have their own separate entity at this point, so we need to figure out which is the correct identifier and merge all instances of the false one into the correct one.  We will do this by taking the true name for the identifier from DO directly, then see wheather the node's ID or it's cross-reference maps to the same name as the node's name.

import pandas as pd
from pathlib import Path
from data_tools.wiki import get_curi_xrefs
from data_tools.df_processing import expand_col_on_char, combine_group_cols_on_char

import obonet
import networkx

this_name = '10b_Diseae_Phenotype_Cleanup_preliminary_DOID_cleaning'
out_dir = Path('../2_pipeline').resolve().joinpath(this_name, 'out')

out_dir.mkdir(parents=True, exist_ok=True)

nw_name = '09_GAUSS_Geneset_to_Phenotype'
nw_dir = Path('../2_pipeline').resolve().joinpath(nw_name, 'out')

nodes = pd.read_csv(nw_dir.joinpath('nodes.csv'), dtype=str)
edges = pd.read_csv(nw_dir.joinpath('edges.csv'), dtype=str)

doid = 'https://github.com/DiseaseOntology/HumanDiseaseOntology/raw/master/src/ontology/doid.obo'
graph = obonet.read_obo(doid)
len(graph)

id_to_name = {id_: data.get('name') for id_, data in graph.nodes(data=True)}

nwid_to_name = nodes.set_index('id')['name'].to_dict()

diseases = nodes.query('label == "Disease"')

doid_xref = get_curi_xrefs(diseases, 'DOID')
doid_xref['id_name'] = doid_xref['id'].map(id_to_name)
doid_xref['xref_name'] = doid_xref['xrefs'].map(id_to_name)
doid_xref['nw_name'] = doid_xref['id'].map(nwid_to_name)

mismatch_doid_xref = doid_xref[doid_xref['id'] != doid_xref['xrefs']].copy()

disease_ids = diseases['id'].unique()

len(mismatch_doid_xref)

mismatch_doid_xref.query('xrefs in @disease_ids')

These are all duplicated ids that should be merged...

keep_id = mismatch_doid_xref[mismatch_doid_xref['id_name'] == mismatch_doid_xref['nw_name']]['id'].unique()
mismatch_doid_xref.query('id in @keep_id and xrefs in @keep_id')

fix_disease = diseases.query('id in @keep_id')
fix_idx = fix_disease.index

fix_disease = expand_col_on_char(fix_disease, 'xrefs', '|')

fix_disease.query('xrefs in @keep_id and id != xrefs')

drop_idx = fix_disease.query('xrefs in @keep_id and id != xrefs').index
fix_disease = fix_disease.drop(drop_idx)

fix_disease = combine_group_cols_on_char(fix_disease, ['id'], ['xrefs'], sort=True, prog=False)

diseases = diseases.drop(fix_idx)
diseases = pd.concat([diseases, fix_disease], sort=False, ignore_index=True)

doid_xref = get_curi_xrefs(diseases, 'DOID')
doid_xref['id_name'] = doid_xref['id'].map(id_to_name)
doid_xref['xref_name'] = doid_xref['xrefs'].map(id_to_name)
doid_xref['nw_name'] = doid_xref['id'].map(nwid_to_name)

mismatch_doid_xref = doid_xref[doid_xref['id'] != doid_xref['xrefs']].copy()

either = mismatch_doid_xref[mismatch_doid_xref['id_name'].isnull() & mismatch_doid_xref['xref_name'].isnull()]['id'].unique()

mismatch_doid_xref.query('id in @either')

set(either)

keep_id = mismatch_doid_xref[~mismatch_doid_xref['id_name'].isnull() & mismatch_doid_xref['xref_name'].isnull()]['id'].unique()
keep_xref = mismatch_doid_xref[mismatch_doid_xref['id_name'].isnull() & ~mismatch_doid_xref['xref_name'].isnull()]['xrefs'].unique()

set(keep_id) == set(keep_xref)

mismatch_doid_xref.query('xrefs in @keep_id')

old_to_new = mismatch_doid_xref.query('id in @keep_id').set_index('xrefs')['id'].to_dict()

old_to_new = {**mismatch_doid_xref.query('xrefs in @keep_id').set_index('id')['xrefs'].to_dict(), **old_to_new}
old_to_new = {**{_id:_id for _id in keep_id}, **old_to_new}

mismatch_doid_xref['new_id'] = mismatch_doid_xref['id'].map(old_to_new)
mismatch_doid_xref['new_id'] = mismatch_doid_xref['new_id'].fillna(mismatch_doid_xref['xrefs'].map(old_to_new))

unmap_subset = mismatch_doid_xref[mismatch_doid_xref['new_id'].isnull()]
unmap_subset

len(unmap_subset)

unmap_subset.drop_duplicates(subset='nw_name', keep=False)

These two are very close, we'll do them by hand

old_to_new = {**{'DOID:11099': 'DOID:1709', 'DOID:1709': 'DOID:1709',
                 'DOID:0050523': 'DOID:0050523', 'DOID:5603':'DOID:0050523'}, **old_to_new}

mismatch_doid_xref['new_id'] = mismatch_doid_xref['id'].map(old_to_new)
mismatch_doid_xref['new_id'] = mismatch_doid_xref['new_id'].fillna(mismatch_doid_xref['xrefs'].map(old_to_new))
unmap_subset = mismatch_doid_xref[mismatch_doid_xref['new_id'].isnull()]
unmap_subset

dup_ids = unmap_subset.drop_duplicates(subset='nw_name')['id'].unique()
dup_xref = unmap_subset.drop_duplicates(subset='nw_name')['xrefs'].unique()

len(set(unmap_subset[['id', 'xrefs']].stack()))

len(unmap_subset.query('id in @dup_xref'))

old_to_new = {**unmap_subset.query('id in @dup_ids').set_index('xrefs')['id'].to_dict(), **old_to_new}
old_to_new = {**{_id:_id for _id in dup_ids}, **old_to_new}

mismatch_doid_xref['new_id'] = mismatch_doid_xref['id'].map(old_to_new)
mismatch_doid_xref['new_id'] = mismatch_doid_xref['new_id'].fillna(mismatch_doid_xref['xrefs'].map(old_to_new))
mismatch_doid_xref.count()

len(mismatch_doid_xref['new_id'].drop_duplicates())

to_query = mismatch_doid_xref['id'].unique()
to_fix = diseases.query('id in @to_query')
fix_idx = to_fix.index

diseases = diseases.drop(fix_idx)

to_fix = expand_col_on_char(to_fix, 'xrefs', '|')
to_fix['id'] = to_fix['id'].map(old_to_new)
fixed = combine_group_cols_on_char(to_fix, ['id'], ['xrefs'], sort=True, prog=False)

len(fixed)

diseases = pd.concat([diseases, fixed], sort=False, ignore_index=True)

fixed_ids = list(old_to_new.keys())

fix_edges = edges.query('start_id in @fixed_ids or end_id in @fixed_ids').copy()
fix_idx = fix_edges.index

edges = edges.drop(fix_idx)

fix_edges['start_id'] = fix_edges['start_id'].map(old_to_new).fillna(fix_edges['start_id'])
fix_edges['end_id'] = fix_edges['end_id'].map(old_to_new).fillna(fix_edges['end_id'])

fix_edges = combine_group_cols_on_char(fix_edges, ['start_id', 'end_id', 'type'], sort=True, prog=False)

edges = pd.concat([edges, fix_edges], ignore_index=True, sort=False)

pd.DataFrame({'id_old': old_to_new.keys(), 'id': old_to_new.values()}).to_csv(out_dir.joinpath('old_to_new_map.csv'), index=False)

nodes = pd.concat([nodes.query('label != "Disease"'), diseases], sort=False, ignore_index=True)

nodes.to_csv(out_dir.joinpath('nodes.csv'), index=False)
edges.to_csv(out_dir.joinpath('edges.csv'), index=False)
