==> 1_code/11_Incorporation_of_Phenotypes_HPO.ipynb <==
%matplotlib inline

import pandas as pd
from pathlib import Path
from biothings_client import get_client

from metapaths.tools import obo_tools as ot
from metapaths.tools.processing import expand_col_on_char, expand_split_col
from hetnet_ml.src import graph_tools as gt


data_dir = Path('../2_pipeline/00_download_data/out/').resolve()
this_name = '11_Incorporation_of_Phenotypes_HPO'
out_dir = Path('../2_pipeline/').joinpath(this_name).joinpath('out').resolve()

if not out_dir.exists():
    out_dir.mkdir(parents=True)

hp_nodes = ot.get_ontology_nodes(data_dir.joinpath('hpo.obo'))
hp_struct = ot.get_ontology_edges(data_dir.joinpath('hpo.obo'))

tab_cols = ['db', 'db_object_id', 'db_name', 'qualifier', 'hp_id', 'db_reference', 'evidence_code', 'onset_modifier', 
            'frequency', 'sex', 'modifier', 'aspect', 'date_created', 'assigned_by']

hp_anno = pd.read_csv(data_dir.joinpath('phenotype_annotation.tab'), header=None, 
                       sep='\t', dtype=str, names=tab_cols)
hp_gene = pd.read_csv(data_dir.joinpath('ALL_SOURCES_ALL_FREQUENCIES_genes_to_phenotype.txt'), 
                       sep='\t', header=None, comment='#', dtype=str,
                       names=['gene_id', 'gene_symbol', 'hp_name', 'hp_id'])

hp_name_map = hp_nodes.set_index('id')['name'].to_dict()
hp_anno['hp_name'] = hp_anno['hp_id'].map(hp_name_map)

len(hp_anno), len(hp_gene)

hp_anno.head(2)

hp_gene.head(2)

## Examine Genes

Genes are using entrez gene_id, so looking for overlap to the current network should be stratightforward

hp_genes = hp_gene['gene_id'].unique()

nodes = gt.remove_colons(pd.read_csv('../2_pipeline/07_Filter_non_human_annotations/out/nodes_filt.csv', dtype=str))
nodes.query('id in @hp_genes')

hp_gene['gene_id'].nunique()

node_ids = set(nodes['id'])
set(hp_gene['gene_id']) - node_ids

Only 1 gene is missing... Don't think we'll add it, but let's at least look at what it is

mg = get_client('gene')
mg.getgene(101928376, fields=['name', 'symbol', 'uniprot'])

## Now HP to Disease Links

HP has OMIM ids for diseases... luckily our CTD network already has xrefs to OMIM for many diseases, so we will use these to merge concepts.

hp_anno.query('db == "OMIM"')['db_object_id'].nunique()

hp_anno['disease_id'] = hp_anno['db'] + ':' + hp_anno['db_object_id']

nodes.query('label == "Disease"')['id'].nunique()

mesh_to_omim = (expand_col_on_char(nodes.query('label == "Disease"'), 'alt_disease_ids', '|')
                    .dropna(subset=['alt_disease_ids'])
                    .set_index('id')['alt_disease_ids'].to_dict())

len(set(hp_anno.query('db == "OMIM"')['disease_id']) - set(mesh_to_omim.values()))

missing_diseases = set(hp_anno.query('db == "OMIM"')['disease_id']) - set(mesh_to_omim.values())
hp_anno.query('disease_id in @missing_diseases')['disease_id'].nunique()

found_diseases = set(hp_anno.query('db == "OMIM"')['disease_id']) & set(mesh_to_omim.values())
hp_to_dis = hp_anno.query('disease_id in @found_diseases').copy()

print(hp_to_dis['disease_id'].nunique())
print(len(hp_to_dis))

hp_to_dis['evidence_code'].value_counts()

# Tons of edges, so remove the electronic annotations
hp_to_dis = hp_to_dis.query('evidence_code != "IEA"').copy()
hp_to_dis = hp_to_dis.drop_duplicates(subset=['disease_id', 'hp_id'])
hp_to_dis.head(4)

So a lot of HPO Items have xrefs to either mesh or omim... thos that are already classified as dieases in our network are not of interest... So we will compare the Xrefs of HPO ids to our network and select those which do not have an xref.

hp_xrefs = expand_split_col(hp_nodes['xref'].dropna().str.split('|'))['xref'].str.replace('MSH:', 'MESH:').unique()

hp_to_xref = expand_split_col(hp_nodes.set_index('id')['xref'].dropna().str.split('|')).rename(columns={'old_idx': 'hp_id'})
(hp_to_xref[hp_to_xref['xref'].str.contains('MSH:')]['hp_id'].value_counts() > 1).sum()

hp_to_xref['hp_id'].nunique()

disease_ids = nodes.query('label == "Disease"')['id'].unique()

hp_to_xref['xref'] = hp_to_xref['xref'].str.replace('MSH:', 'MESH:')
hp_disease_ids = hp_to_xref.query('xref in @disease_ids')['hp_id'].unique()
non_disease_hp = hp_to_xref.query('xref not in @disease_ids').copy()
non_disease_hp['name'] = non_disease_hp['hp_id'].map(hp_name_map)
len(non_disease_hp[non_disease_hp['xref'].str.contains('MESH:')])

hp_to_dis = hp_to_dis.query('hp_id not in @hp_disease_ids').copy()
print(len(hp_to_dis))
print(hp_to_dis['hp_id'].nunique())
print(hp_to_dis['disease_id'].nunique())

hp_to_dis.sample(10)

hp_to_keep = hp_to_dis['hp_id'].unique()

So these are the HPO ids that we'll keep for our network... 
Not all the Gene to HP ids will be contained in this set so we need to filter a bit

hp_gene.query('hp_id in @hp_to_keep and gene_id in @node_ids')['gene_id'].nunique()

hp_gene.query('hp_id in @hp_to_keep and gene_id in @node_ids')['hp_id'].nunique()

len(hp_gene.query('hp_id in @hp_to_keep and gene_id in @node_ids'))

## Aggregate the new edges and save to disk

hp_to_dis_edges = hp_to_dis.rename(columns={'disease_id': 'start_id', 'hp_id': 'end_id'})
hp_to_dis_edges['start_id'] = hp_to_dis_edges['start_id'].map({v:k for k, v in mesh_to_omim.items()})
hp_to_dis_edges['type'] = 'presents_DpPH'

hp_to_dis_edges[['start_id', 'end_id', 'type']].head(2)

hp_to_gene_edges = hp_gene.query('hp_id in @hp_to_keep and gene_id in @node_ids').copy()
hp_to_gene_edges = hp_to_gene_edges.rename(columns={'hp_id': 'start_id', 'gene_id': 'end_id'})
hp_to_gene_edges['type'] = 'associated_with_PHawG'

hp_to_gene_edges[['start_id', 'end_id', 'type']].head(2)

all_edges = pd.concat([hp_to_dis_edges[['start_id', 'end_id', 'type']], 
                       hp_to_gene_edges[['start_id', 'end_id', 'type']]], ignore_index=True)
all_edges.head(2)

edge_ids = set(all_edges['start_id']).union(set(all_edges['end_id']))
hp_nodes_filt = hp_nodes.query('id in @edge_ids')[['id', 'name']]
hp_nodes_filt['label'] = 'Phenotype'

all_edges.to_csv(out_dir.joinpath('edges.csv'), index=False)
hp_nodes_filt.to_csv(out_dir.joinpath('nodes.csv'), index=False)
