==> 1_code/12_Taxonomic_cause_of_disease.ipynb <==
import pandas as pd
from pathlib import Path
from biothings_client import get_client
from metapaths.tools import obo_tools as ot
from wikidataintegrator.wdi_core import WDItemEngine
from tools.processing import strip_chars, expand_col_on_char, combine_group_rows_on_char
from tools.biothings_processing import get_dotval

name = '12_Taxonomic_cause_of_disease'
out_dir = (Path('../2_pipeline').joinpath(name).joinpath('out')).resolve()
if not out_dir.exists():
    out_dir.mkdir(parents=True)

tax_dis = pd.read_csv('../0_data/manual/infectious-taxa.csv')

tax_dis.head(2)

# Fix weird formatting issues
for c in tax_dis:
    tax_dis[c] = tax_dis[c].str.strip()

data_dir = Path('../2_pipeline/00_download_data/out/').resolve()
tax_items = ot.get_ontology_nodes(data_dir.joinpath('ncbitaxon.obo'))

tax_items.head(2)

tax_items['strp_name'] = tax_items['name'].str.lower().apply(strip_chars)
name_to_id = tax_items.set_index('strp_name')['id'].to_dict()

tax_dis['strp_name'] = tax_dis['infectious_agent'].str.lower()
tax_dis['strp_name'] = tax_dis['strp_name'].str.split(' \(', expand=True)[0]
tax_dis['strp_name'] = tax_dis['strp_name'].apply(strip_chars)

tax_dis['tax_id'] = tax_dis['strp_name'].map(name_to_id)
tax_dis.count()

idx = tax_dis['tax_id'].isnull()
print(idx.sum())
tax_dis[idx]

# Disease mappings

nodes = pd.read_csv('../2_pipeline/07_Filter_non_human_annotations/out/nodes_all.csv', dtype=str)

disease_name_map = nodes[nodes[':LABEL'] == 'Disease'].set_index('name')['identifier:ID'].to_dict()

strp_d_name = {strip_chars(k).lower(): v for k, v in disease_name_map.items()}

tax_dis['strp_name'] = tax_dis['common_name'].str.lower()
tax_dis['strp_name'] = tax_dis['strp_name'].str.split(' \(', expand=True)[0]
tax_dis['strp_name'] = tax_dis['strp_name'].apply(strip_chars)

tax_dis['mesh_id'] = tax_dis['strp_name'].map(strp_d_name)

tax_dis_found = tax_dis.dropna(subset=['tax_id', 'mesh_id']).copy()
tax_dis_notfound = tax_dis[tax_dis['tax_id'].isnull() | tax_dis['mesh_id'].isnull()].copy()

tax_dis_found.sample()

wd_query = """SELECT Distinct ?dis ?disLabel ?mesh_id ?doid ?taxLabel ?tax_id
WHERE 
{
  ?dis wdt:P31 wd:Q12136.
  OPTIONAL{?dis wdt:P486 ?mesh_id}.
  OPTIONAL{?dis wdt:P699 ?doid}.
  {?dis wdt:P828 ?tax} UNION {?tax wdt:P1542 ?dis}.
  ?tax wdt:P685 ?tax_id.
  SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGUAGE],en". }
}"""

qres = WDItemEngine.execute_sparql_query(wd_query, as_dataframe=True)
qres['mesh_id'] = 'MESH:' + qres['mesh_id']
qres['tax_id'] = 'NCBITaxon:' + qres['tax_id']
qres = qres.drop('dis', axis=1)
qres = qres.dropna(subset=['doid', 'mesh_id'], how='all')
qres = qres.rename(columns={'disLabel': 'common_name', 'taxLabel': 'infectious_agent'})
qres.head(5)

len(qres)

tax_dis_found = pd.concat([qres, tax_dis_found], sort=False).drop_duplicates(subset=['mesh_id', 'tax_id'])
len(tax_dis_found)

# biothings

def remove_paren(s):
    return s.split('(')[0].strip()

def get_inner_paren(s):
    start = s.index('(')+1
    end = s.index(')')
    return s[start:end]

def pipe_combine_paren(s):
    if '(' not in s:
        return s
    before_paren = remove_paren(s)
    in_paren = get_inner_paren(s)
    return before_paren + '|' + in_paren
        

# Split Parenthesis on common names and infectious agents
tax_dis_notfound['common_name'] = tax_dis_notfound['common_name'].apply(pipe_combine_paren)
tax_dis_notfound = expand_col_on_char(tax_dis_notfound, 'common_name', '|')

tax_dis_notfound['infectious_agent'] = tax_dis_notfound['infectious_agent'].apply(pipe_combine_paren)
tax_dis_notfound = expand_col_on_char(tax_dis_notfound, 'infectious_agent', '|')

# Lowercase for queries
tax_dis_notfound['dis_qname'] = tax_dis_notfound['common_name'].str.lower()
tax_dis_notfound['tax_qname'] = tax_dis_notfound['infectious_agent'].str.lower()

# Get the values to query
dis_to_query = tax_dis_notfound[tax_dis_notfound['mesh_id'].isnull()]['dis_qname'].unique().tolist()
tax_to_query = tax_dis_notfound[tax_dis_notfound['tax_id'].isnull()]['tax_qname'].unique().tolist()

# Special formatting for biothings_client bug....
dis_to_query = ['"{}"'.format(d) for d in dis_to_query]
dis_to_query[:4]

md = get_client('disease')

# See potential fields for names...
[k for k in md.get_fields() if 'name' in k or 'label' in k]

scopes=['mondo.label', 'hpo.disease_name','disgenet.xrefs.disease_name']
fields=['ctd.mesh','ctd.omim','mondo.xrefs','disgenet.xrefs', 'hpo.omim']+scopes

res = md.querymany(dis_to_query, scopes=scopes, fields=fields)
exact_mondo_res = [r for r in res if strip_chars(r['query']) == 
                   strip_chars(r.get('mondo', dict()).get('label', '').lower())]

exact_hpo_res = [r for r in res if strip_chars(r['query']) == 
                 strip_chars(r.get('hpo', dict()).get('disease_name', '').lower())]

exact_disgenet_res = []

for r in res:
    name = get_dotval(r, 'disgenet.xrefs.disease_name')

    if type(name) == list:
        names = [strip_chars(n).lower() for n in name]
        if strip_chars(r['query']).lower() in names:
            exact_disgenet_res.append(r)

    elif pd.isnull(name):
        continue 

    else:
        if strip_chars(r['query']).lower() == strip_chars(name).lower():
            exact_disgenet_res.append(r)


names = []
mesh_ids = []
doids = []
omim_ids = []

for r in exact_mondo_res:
    mesh = get_dotval(r, 'mondo.xrefs.mesh')
    doid = get_dotval(r, 'mondo.xrefs.doid')
    omim = get_dotval(r, 'mondo.xrefs.omim')

    # Look at CTD for things that were missed
    if pd.isnull(mesh):
        mesh = get_dotval(r, 'ctd.mesh')
    if pd.isnull(omim):
        omim = get_dotval(r, 'ctd.omim')

    names.append(r['query'])
    mesh_ids.append(mesh)
    doids.append(doid)
    omim_ids.append(omim)
    
mondo_res = pd.DataFrame({'name': names, 'mesh_id': mesh_ids, 'doid': doids, 'omim_id': omim_ids})
mondo_res['mesh_id'] = 'MESH:' + mondo_res['mesh_id']
mondo_res['omim_id'] = 'OMIM:' + mondo_res['omim_id']
mondo_res.head(5)

mondo_res.dropna(subset=['omim_id'])

names = []
mesh_ids = []
omim_ids = []

for r in exact_hpo_res:
    omim = get_dotval(r, 'hpo.omim')    
    mesh = get_dotval(r, 'ctd.mesh')
    
    if pd.isnull(omim):
        omim = get_dotval(r, 'ctd.omim')
    
    names.append(r['query'])
    mesh_ids.append(mesh)
    omim_ids.append(omim)
    
hpo_res = pd.DataFrame({'name': names, 'mesh_id': mesh_ids, 'omim_id': omim_ids})
hpo_res['mesh_id'] = 'MESH:' + hpo_res['mesh_id']
hpo_res['omim_id'] = 'OMIM:' + hpo_res['omim_id']
hpo_res.head(5)

def add_prefix_list_or_string(items, prefix):
    if type(items) == list:
        return '|'.join([prefix+i for i in set(items)])
    elif type(items) == str:
        return prefix + items
    else:
        return items

names = []
mesh_ids = []
doids = []
omim_ids = []

for r in exact_disgenet_res:
    mesh = get_dotval(r, 'disgenet.xrefs.mesh')
    doid = get_dotval(r, 'disgenet.xrefs.doid')
    omim = get_dotval(r, 'disgenet.xrefs.omim', list_ok=False)
    
    # Look at CTD for things that were missed
    if type(mesh) != list and pd.isnull(mesh):
        mesh = get_dotval(r, 'ctd.mesh')
    if pd.isnull(omim):
        omim = get_dotval(r, 'ctd.omim')

    names.append(r['query'])
    
    mesh_ids.append(add_prefix_list_or_string(mesh, 'MESH:'))
    doids.append(add_prefix_list_or_string(doid, 'DOID:'))
    omim_ids.append(omim)
    
disgenet_res = pd.DataFrame({'name': names, 'mesh_id': mesh_ids, 'doid': doids, 'omim_id': omim_ids})
disgenet_res['omim_id'] = 'OMIM:' + disgenet_res['omim_id']
disgenet_res

found_dis = pd.concat([mondo_res, hpo_res, disgenet_res], sort=False, ignore_index=False)

found_dis = combine_group_rows_on_char(found_dis, 'name').replace('', float('nan'))

found_dis

### Query MyTaxon.info

tax_to_query = ['"{}"'.format(d) for d in tax_to_query]
tax_to_query[:4]

mt = get_client('taxon')

[k for k in mt.get_fields() if 'name' in k or 'label' in k]

scopes = [k for k in mt.get_fields() if 'name' in k]
fields = ['taxid']+scopes

res = mt.querymany(tax_to_query, scopes=scopes, fields=fields, size=100000)

exact_sname = [r for r in res if strip_chars(r.get('scientific_name', '')).lower() == strip_chars(r['query'])]
exact_cname = [r for r in res if strip_chars(r.get('common_name', '').lower()) == strip_chars(r['query'])]
exact_uname = [r for r in res if strip_chars(r.get('uniprot_name', '')).lower() == strip_chars(r['query'])]
exact_gname = [r for r in res if strip_chars(r.get('genbank_common_name', '').lower()) == strip_chars(r['query'])]

tax_map = dict()
for item in exact_sname + exact_cname + exact_uname + exact_gname:
    tax_map[item['query']] = 'NCBITaxon:' + str(item['taxid'])

tax_dis_notfound['tax_id'] = tax_dis_notfound['tax_id'].fillna(tax_dis_notfound['tax_qname'].map(tax_map))

tax_dis_notfound = tax_dis_notfound.merge(found_dis, left_on='dis_qname', right_on='name', how='left')
tax_dis_notfound.head(2)

tax_dis_notfound['mesh_id_x'] = tax_dis_notfound['mesh_id_x'].fillna(tax_dis_notfound['mesh_id_y'])
tax_dis_notfound = tax_dis_notfound.rename(columns={'mesh_id_x': 'mesh_id'}).drop('mesh_id_y', axis=1)
tax_dis_notfound.head(2)

tax_dis_notfound.count()

null_dis = tax_dis_notfound['doid'].isnull() & \
           tax_dis_notfound['mesh_id'].isnull() & \
           tax_dis_notfound['omim_id'].isnull() 
null_tax = tax_dis_notfound['tax_id'].isnull()

missing_info = tax_dis_notfound[null_dis | null_tax]

len(missing_info)

tax_dis_notfound['dis_id'] = tax_dis_notfound['mesh_id'].fillna(tax_dis_notfound['omim_id']).fillna(tax_dis_notfound['doid'])
tax_dis_notfound = tax_dis_notfound.dropna(subset=['dis_id', 'tax_id'])

tax_dis_notfound = expand_col_on_char(tax_dis_notfound, 'dis_id', '|')

tax_dis_notfound = tax_dis_notfound.drop_duplicates(subset=['tax_id', 'dis_id'])

tax_dis_found['dis_id'] = tax_dis_found['mesh_id'].fillna(tax_dis_found['doid']) 

tax_dis_final = pd.concat([tax_dis_found, tax_dis_notfound], sort=False, ignore_index=True).drop_duplicates(subset=['tax_id', 'dis_id'])
len(tax_dis_final)

tax_name_map = tax_items.set_index('id')['name']
tax_nodes = pd.Series(tax_dis_final['tax_id'].unique(), name='id').to_frame()
tax_nodes['name'] = tax_nodes['id'].map(tax_name_map)
tax_nodes['label'] = 'Taxonomy'

tax_nodes.head()

dis_ids = set(disease_name_map.values())
dis_nodes = (tax_dis_final.query('dis_id not in @dis_ids')[['dis_id', 'common_name']]
                .rename(columns={'dis_id': 'id', 'common_name': 'name'}).drop_duplicates())
dis_nodes['name'] = dis_nodes['name'].str.title()
dis_nodes['label'] = 'Disease'
dis_nodes.head()

edges = tax_dis_final.rename(columns={'tax_id': 'start_id', 'dis_id': 'end_id'})[['start_id', 'end_id']]
edges['type'] = 'Causes_TcD'
edges.head(10)

nodes = pd.concat([tax_nodes, dis_nodes])
nodes.to_csv(out_dir.joinpath('new_nodes.csv'), index=False)
edges.to_csv(out_dir.joinpath('edges.csv'), index=False)
