==> 1_code/12a_Preprocessing-Semmantic_Compression.ipynb <==
# Semmantic Compression

Based on metaedges and counts (created in the previous notebook), we made a semmantic compression map. This will allow us to map similarly typed edges together to create a more limited semmantic vocabularly.

%matplotlib inline
import pandas as pd
import seaborn as sns
from pathlib import Path
import matplotlib.pyplot as plt

from data_tools.plotting import count_plot_h
from data_tools import combine_nodes_and_edges
from data_tools.df_processing import combine_group_cols_on_char, expand_col_on_char
from data_tools.graphs import map_edge_types_from_file, add_abbrevs, get_core_network

prev_dir = Path('../2_pipeline/11_Reactions_and_Regulation/out').resolve()
data_dir = Path('../0_data/manual/').resolve()

nodes = pd.read_csv(prev_dir.joinpath('nodes.csv'), dtype=str)
edges = pd.read_csv(prev_dir.joinpath('edges.csv'), dtype=str)

sem_comp_map = pd.read_csv(data_dir.joinpath('semantic_compression_map.csv'))

sem_comp_map.dtypes

sem_comp_map

len(sem_comp_map[['start_label', 'new_type', 'end_label']].dropna().drop_duplicates())

So we are reducing our total number of metaedges from 605 to 235, about a 60% reduction

The mapping function is in-place, so lets do a quick before and after comparison to make sure things are mapped

len(edges), len(edges.drop_duplicates(subset=['start_id', 'end_id', 'type']))

map_edge_types_from_file(edges, sem_comp_map, nodes=nodes)

len(edges), len(edges.drop_duplicates(subset=['start_id', 'end_id', 'type']))

edges.head(10)

edges = combine_group_cols_on_char(edges, ['start_id', 'end_id', 'type'], sort=True)

As some edge types were dropped, there may be fewer nodes

edge_ids = edges[['start_id', 'end_id']].stack().unique()

len(nodes), len(nodes.query('id in @edge_ids'))

### Final Checks

Don't need to hold onto legacy nodes anymore... only keep those in the network

nodes = nodes.query('id in @edge_ids')

#### Ensure no duplicated nodes

nodes[nodes['id'].duplicated(keep=False)]

nodes = nodes.drop_duplicates(subset=['id'], keep='first')

#### Make sure all nodes have a label

nodes[nodes['label'].isnull()]

All are diseases

disease_idx = nodes[nodes['label'].isnull()].index
nodes.loc[disease_idx, 'label'] = 'Disease'

nodes.loc[disease_idx].head()

### Save results

this_name = '12a_Preprocessing-Semmantic_Compression'

out_dir = Path('../2_pipeline/').joinpath(this_name).joinpath('out').resolve()

out_dir.mkdir(parents=True, exist_ok=True)

nodes.to_csv(out_dir.joinpath('nodes.csv'), index=False)
edges.to_csv(out_dir.joinpath('edges.csv'), index=False)
