==> 1_code/12d_Preprocessing-Abbreviations_and_Biolink.ipynb <==
# Add edge semantic abbrevations and map to Biolink

Abbreviations are needed for some of the machine learning parts, to distinguish edge and node semanitcs in the hetnet. We will add them now.  We will also map node semantic types to the Biolink model

%matplotlib inline
import pandas as pd
import seaborn as sns
from pathlib import Path
import matplotlib.pyplot as plt

from data_tools.plotting import count_plot_h
from data_tools import combine_nodes_and_edges
from data_tools.df_processing import combine_group_cols_on_char, expand_col_on_char
from data_tools.graphs import map_edge_types_from_file, add_abbrevs, get_core_network

prev_dir = Path('../2_pipeline/12c_Preprocessing-Pruning/out').resolve()
data_dir = Path('../0_data/manual/').resolve()

nodes = pd.read_csv(prev_dir.joinpath('nodes.csv'), dtype=str)
edges = pd.read_csv(prev_dir.joinpath('edges.csv'), dtype=str)

## Add Abbreviations

biolink_map = pd.read_csv(data_dir.joinpath('hetnet_to_biolink.csv'))
sem_info = pd.read_csv(data_dir.joinpath('edge_semtypes.csv'))

sem_info

rev_map = sem_info.set_index('fwd_edge')['rev_edge'].to_dict()
directed_map = sem_info.set_index('fwd_edge')['directed'].to_dict()

node_abbrev_map = biolink_map.set_index('hetnet')['abbrev'].to_dict()
edge_abbrev_map = sem_info.set_index('fwd_edge')['abbrev'].to_dict()

%%time
edges = add_abbrevs(nodes, edges, {**node_abbrev_map, **edge_abbrev_map}, directed_map)
edges.head(2)

edges['type_no_abbv'] = edges['type']
edges['type'] = edges['type'] + '_' + edges['abbrev']
edges.head(2)

#### Save network

edges['abbrev'].nunique()

this_name = '12d_Preprocessing-Abbreviations_and_Biolink'

out_dir = Path('../2_pipeline/').joinpath(this_name).joinpath('out').resolve()

out_dir.mkdir(parents=True, exist_ok=True)

nodes.to_csv(out_dir.joinpath('nodes.csv'), index=False)
edges.to_csv(out_dir.joinpath('edges.csv'), index=False)

# Map to biolink

nodes_biolink = nodes.copy()
edges_biolink = edges.copy()

label_to_biolink = biolink_map.set_index('hetnet')['bl_hierarchy_mapping'].to_dict()
node_abbrev_map = biolink_map.set_index('bl_hierarchy_mapping')['hie_abbrev'].to_dict()

nodes_biolink['label'] = nodes_biolink['label'].map(label_to_biolink)

# reset the edge semmantics
edges_biolink['type'] = edges_biolink['type_no_abbv']

# Add in the new abbrevations
edges_biolink = add_abbrevs(nodes_biolink, edges_biolink, {**node_abbrev_map, **edge_abbrev_map}, directed_map)
edges_biolink.head(2)

edges_biolink['abbrev'].nunique()

edges_biolink['type_no_abbv'] = edges_biolink['type']
edges_biolink['type'] = edges_biolink['type'] + '_' + edges_biolink['abbrev']

edges_biolink.head(2)

### Ensure Biolink Mapping hasn't resulted in Reverse Types

For example, if we had `Gene` `part_of` `Cellular Component` and `Cellular Component` `has_part` `Protein`,  Mapping to biolkink would result in `AnatomicalEntity` `has_part` `MacromolecularMachine` and `MacromolecularMachine` `part_of` `AnatomicalEntity`, which are essentially the same edges just reversed.  These need to be normalized if identified.

combo = combine_nodes_and_edges(nodes_biolink, edges_biolink)
combo['rev_type'] = combo['type_no_abbv'].map(rev_map)

# Don't need to do these operations on the entire set of edges, just one of each type is ok
meta_edges = combo.drop_duplicates(subset=['start_label', 'type_no_abbv', 'end_label']).copy()

meta_edges['fwd_tup'] = meta_edges[['start_label', 'type_no_abbv', 'end_label']].apply(tuple, axis=1)
meta_edges['rev_tup'] = meta_edges[['end_label', 'rev_type', 'start_label']].apply(tuple, axis=1)

overlaps = list(set(meta_edges['fwd_tup']) & set(meta_edges['rev_tup']))
overlaps

Empty list is exactly what we were looking for... Perfect!

## Save out

'{:,}'.format(len(nodes_biolink))

'{:,}'.format(len(edges_biolink))

ex = expand_col_on_char(edges_biolink, 'source', '|')

ex['source'].value_counts()

nodes_biolink.to_csv(out_dir.joinpath('nodes_biolink.csv'), index=False)
edges_biolink.to_csv(out_dir.joinpath('edges_biolink.csv'), index=False)
